Move code from BatchFixAllProvider to FixAllContext.
