+++ src/Workspaces/Core/Portable/CodeFixes/FixAllOccurrences/BatchFixAllProvider.cs
@@ -31,12 +31,12 @@ namespace Microsoft.CodeAnalysis.CodeFixes
         {
             if (fixAllContext.Document != null)
             {
-                var documentsAndDiagnosticsToFixMap = await GetDocumentDiagnosticsToFixAsync(fixAllContext).ConfigureAwait(false);
+                var documentsAndDiagnosticsToFixMap = await fixAllContext.GetDocumentDiagnosticsToFixAsync().ConfigureAwait(false);
                 return await GetFixAsync(documentsAndDiagnosticsToFixMap, fixAllContext).ConfigureAwait(false);
             }
             else
             {
-                var projectsAndDiagnosticsToFixMap = await GetProjectDiagnosticsToFixAsync(fixAllContext).ConfigureAwait(false);
+                var projectsAndDiagnosticsToFixMap = await fixAllContext.GetProjectDiagnosticsToFixAsync().ConfigureAwait(false);
                 return await GetFixAsync(projectsAndDiagnosticsToFixMap, fixAllContext).ConfigureAwait(false);
             }
         }
@@ -237,165 +237,6 @@ namespace Microsoft.CodeAnalysis.CodeFixes
 
                 default:
                     throw ExceptionUtilities.Unreachable;
-            }
-        }
-
-        internal override async Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(FixAllContext fixAllContext)
-        {
-            using (Logger.LogBlock(FunctionId.CodeFixes_FixAllOccurrencesComputation_Diagnostics, fixAllContext.CancellationToken))
-            {
-                var allDiagnostics = ImmutableArray<Diagnostic>.Empty;
-                var projectsToFix = ImmutableArray<Project>.Empty;
-
-                var document = fixAllContext.Document;
-                var project = fixAllContext.Project;
-                var generatedCodeServices = project.Solution.Workspace.Services.GetService<IGeneratedCodeRecognitionService>();
-
-                switch (fixAllContext.Scope)
-                {
-                    case FixAllScope.Document:
-                        if (document != null && !generatedCodeServices.IsGeneratedCode(document))
-                        {
-                            var documentDiagnostics = await fixAllContext.GetDocumentDiagnosticsAsync(document).ConfigureAwait(false);
-                            var kvp = SpecializedCollections.SingletonEnumerable(KeyValuePair.Create(document, documentDiagnostics));
-                            return ImmutableDictionary.CreateRange(kvp);
-                        }
-
-                        break;
-
-                    case FixAllScope.Project:
-                        projectsToFix = ImmutableArray.Create(project);
-                        allDiagnostics = await fixAllContext.GetAllDiagnosticsAsync(project).ConfigureAwait(false);
-                        break;
-
-                    case FixAllScope.Solution:
-                        projectsToFix = project.Solution.Projects
-                            .Where(p => p.Language == project.Language)
-                            .ToImmutableArray();
-
-                        var diagnostics = new ConcurrentBag<Diagnostic>();
-                        var tasks = new Task[projectsToFix.Length];
-                        for (int i = 0; i < projectsToFix.Length; i++)
-                        {
-                            fixAllContext.CancellationToken.ThrowIfCancellationRequested();
-                            var projectToFix = projectsToFix[i];
-                            tasks[i] = Task.Run(async () =>
-                            {
-                                var projectDiagnostics = await fixAllContext.GetAllDiagnosticsAsync(projectToFix).ConfigureAwait(false);
-                                foreach (var diagnostic in projectDiagnostics)
-                                {
-                                    fixAllContext.CancellationToken.ThrowIfCancellationRequested();
-                                    diagnostics.Add(diagnostic);
-                                }
-                            }, fixAllContext.CancellationToken);
-                        }
-
-                        await Task.WhenAll(tasks).ConfigureAwait(false);
-                        allDiagnostics = allDiagnostics.AddRange(diagnostics);
-                        break;
-                }
-
-                if (allDiagnostics.IsEmpty)
-                {
-                    return ImmutableDictionary<Document, ImmutableArray<Diagnostic>>.Empty;
-                }
-
-                return await GetDocumentDiagnosticsToFixAsync(allDiagnostics, projectsToFix, generatedCodeServices.IsGeneratedCode, fixAllContext.CancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        private async static Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(
-            ImmutableArray<Diagnostic> diagnostics,
-            ImmutableArray<Project> projects,
-            Func<Document, bool> isGeneratedCode, CancellationToken cancellationToken)
-        {
-            var treeToDocumentMap = await GetTreeToDocumentMapAsync(projects, cancellationToken).ConfigureAwait(false);
-
-            var builder = ImmutableDictionary.CreateBuilder<Document, ImmutableArray<Diagnostic>>();
-            foreach (var documentAndDiagnostics in diagnostics.GroupBy(d => GetReportedDocument(d, treeToDocumentMap)))
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-                var document = documentAndDiagnostics.Key;
-                if (!isGeneratedCode(document))
-                {
-                    var diagnosticsForDocument = documentAndDiagnostics.ToImmutableArray();
-                    builder.Add(document, diagnosticsForDocument);
-                }
-            }
-
-            return builder.ToImmutable();
-        }
-
-        private static async Task<ImmutableDictionary<SyntaxTree, Document>> GetTreeToDocumentMapAsync(ImmutableArray<Project> projects, CancellationToken cancellationToken)
-        {
-            var builder = ImmutableDictionary.CreateBuilder<SyntaxTree, Document>();
-            foreach (var project in projects)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-                foreach (var document in project.Documents)
-                {
-                    cancellationToken.ThrowIfCancellationRequested();
-                    var tree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
-                    builder.Add(tree, document);
-                }
-            }
-
-            return builder.ToImmutable();
-        }
-
-        private static Document GetReportedDocument(Diagnostic diagnostic, ImmutableDictionary<SyntaxTree, Document> treeToDocumentsMap)
-        {
-            var tree = diagnostic.Location.SourceTree;
-            if (tree != null)
-            {
-                Document document;
-                if (treeToDocumentsMap.TryGetValue(tree, out document))
-                {
-                    return document;
-                }
-            }
-
-            return null;
-        }
-
-        internal override async Task<ImmutableDictionary<Project, ImmutableArray<Diagnostic>>> GetProjectDiagnosticsToFixAsync(FixAllContext fixAllContext)
-        {
-            using (Logger.LogBlock(FunctionId.CodeFixes_FixAllOccurrencesComputation_Diagnostics, fixAllContext.CancellationToken))
-            {
-                var project = fixAllContext.Project;
-                if (project != null)
-                {
-                    switch (fixAllContext.Scope)
-                    {
-                        case FixAllScope.Project:
-                            var diagnostics = await fixAllContext.GetProjectDiagnosticsAsync(project).ConfigureAwait(false);
-                            var kvp = SpecializedCollections.SingletonEnumerable(KeyValuePair.Create(project, diagnostics));
-                            return ImmutableDictionary.CreateRange(kvp);
-
-                        case FixAllScope.Solution:
-                            var projectsAndDiagnostics = ImmutableDictionary.CreateBuilder<Project, ImmutableArray<Diagnostic>>();
-
-                            var tasks = project.Solution.Projects.Select(async p => new
-                            {
-                                Project = p,
-                                Diagnostics = await fixAllContext.GetProjectDiagnosticsAsync(p).ConfigureAwait(false)
-                            }).ToArray();
-
-                            await Task.WhenAll(tasks).ConfigureAwait(false);
-
-                            foreach (var task in tasks)
-                            {
-                                if (task.Result.Diagnostics.Any())
-                                {
-                                    projectsAndDiagnostics[task.Result.Project] = task.Result.Diagnostics;
-                                }
-                            }
-
-                            return projectsAndDiagnostics.ToImmutable();
-                    }
-                }
-
-                return ImmutableDictionary<Project, ImmutableArray<Diagnostic>>.Empty;
             }
         }
