+++ src/Workspaces/Core/Portable/CodeFixes/FixAllOccurrences/FixAllContext.DiagnosticProvider.cs
@@ -5,6 +5,11 @@ using System.Threading.Tasks;
 using System.Threading;
 using System;
 using System.Collections.Immutable;
+using Microsoft.CodeAnalysis.Internal.Log;
+using Microsoft.CodeAnalysis.GeneratedCodeRecognition;
+using Roslyn.Utilities;
+using System.Linq;
+using System.Collections.Concurrent;
 
 namespace Microsoft.CodeAnalysis.CodeFixes
 {
@@ -36,29 +41,165 @@ namespace Microsoft.CodeAnalysis.CodeFixes
             /// </summary>
             public abstract Task<IEnumerable<Diagnostic>> GetAllDiagnosticsAsync(Project project, CancellationToken cancellationToken);
 
-            internal virtual Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(
-                FixAllProvider batchFixer, FixAllContext context)
+            internal virtual async Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(
+                FixAllContext fixAllContext)
             {
-                return batchFixer.GetDocumentDiagnosticsToFixAsync(context);
+                using (Logger.LogBlock(FunctionId.CodeFixes_FixAllOccurrencesComputation_Diagnostics, fixAllContext.CancellationToken))
+                {
+                    var allDiagnostics = ImmutableArray<Diagnostic>.Empty;
+                    var projectsToFix = ImmutableArray<Project>.Empty;
+
+                    var document = fixAllContext.Document;
+                    var project = fixAllContext.Project;
+                    var generatedCodeServices = project.Solution.Workspace.Services.GetService<IGeneratedCodeRecognitionService>();
+
+                    switch (fixAllContext.Scope)
+                    {
+                        case FixAllScope.Document:
+                            if (document != null && !generatedCodeServices.IsGeneratedCode(document))
+                            {
+                                var documentDiagnostics = await fixAllContext.GetDocumentDiagnosticsAsync(document).ConfigureAwait(false);
+                                var kvp = SpecializedCollections.SingletonEnumerable(KeyValuePair.Create(document, documentDiagnostics));
+                                return ImmutableDictionary.CreateRange(kvp);
+                            }
+
+                            break;
+
+                        case FixAllScope.Project:
+                            projectsToFix = ImmutableArray.Create(project);
+                            allDiagnostics = await fixAllContext.GetAllDiagnosticsAsync(project).ConfigureAwait(false);
+                            break;
+
+                        case FixAllScope.Solution:
+                            projectsToFix = project.Solution.Projects
+                                .Where(p => p.Language == project.Language)
+                                .ToImmutableArray();
+
+                            var diagnostics = new ConcurrentBag<Diagnostic>();
+                            var tasks = new Task[projectsToFix.Length];
+                            for (int i = 0; i < projectsToFix.Length; i++)
+                            {
+                                fixAllContext.CancellationToken.ThrowIfCancellationRequested();
+                                var projectToFix = projectsToFix[i];
+                                tasks[i] = Task.Run(async () =>
+                                {
+                                    var projectDiagnostics = await fixAllContext.GetAllDiagnosticsAsync(projectToFix).ConfigureAwait(false);
+                                    foreach (var diagnostic in projectDiagnostics)
+                                    {
+                                        fixAllContext.CancellationToken.ThrowIfCancellationRequested();
+                                        diagnostics.Add(diagnostic);
+                                    }
+                                }, fixAllContext.CancellationToken);
+                            }
+
+                            await Task.WhenAll(tasks).ConfigureAwait(false);
+                            allDiagnostics = allDiagnostics.AddRange(diagnostics);
+                            break;
+                    }
+
+                    if (allDiagnostics.IsEmpty)
+                    {
+                        return ImmutableDictionary<Document, ImmutableArray<Diagnostic>>.Empty;
+                    }
+
+                    return await GetDocumentDiagnosticsToFixAsync(allDiagnostics, projectsToFix, generatedCodeServices.IsGeneratedCode, fixAllContext.CancellationToken).ConfigureAwait(false);
+                }
             }
 
-            internal virtual Task<ImmutableDictionary<Project, ImmutableArray<Diagnostic>>> GetProjectDiagnosticsToFixAsync(
-                FixAllProvider batchFixer, FixAllContext context)
+            private async static Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(
+                ImmutableArray<Diagnostic> diagnostics,
+                ImmutableArray<Project> projects,
+                Func<Document, bool> isGeneratedCode, CancellationToken cancellationToken)
             {
-                return batchFixer.GetProjectDiagnosticsToFixAsync(context);
+                var treeToDocumentMap = await GetTreeToDocumentMapAsync(projects, cancellationToken).ConfigureAwait(false);
+
+                var builder = ImmutableDictionary.CreateBuilder<Document, ImmutableArray<Diagnostic>>();
+                foreach (var documentAndDiagnostics in diagnostics.GroupBy(d => GetReportedDocument(d, treeToDocumentMap)))
+                {
+                    cancellationToken.ThrowIfCancellationRequested();
+                    var document = documentAndDiagnostics.Key;
+                    if (!isGeneratedCode(document))
+                    {
+                        var diagnosticsForDocument = documentAndDiagnostics.ToImmutableArray();
+                        builder.Add(document, diagnosticsForDocument);
+                    }
+                }
+
+                return builder.ToImmutable();
             }
-        }
 
-        internal Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync(
-            FixAllProvider batchFixer)
-        {
-            return _diagnosticProvider.GetDocumentDiagnosticsToFixAsync(batchFixer, this);
-        }
+            private static async Task<ImmutableDictionary<SyntaxTree, Document>> GetTreeToDocumentMapAsync(ImmutableArray<Project> projects, CancellationToken cancellationToken)
+            {
+                var builder = ImmutableDictionary.CreateBuilder<SyntaxTree, Document>();
+                foreach (var project in projects)
+                {
+                    cancellationToken.ThrowIfCancellationRequested();
+                    foreach (var document in project.Documents)
+                    {
+                        cancellationToken.ThrowIfCancellationRequested();
+                        var tree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
+                        builder.Add(tree, document);
+                    }
+                }
 
-        internal Task<ImmutableDictionary<Project, ImmutableArray<Diagnostic>>> GetProjectDiagnosticsToFixAsync(
-            FixAllProvider batchFixer)
-        {
-            return _diagnosticProvider.GetProjectDiagnosticsToFixAsync(batchFixer, this);
+                return builder.ToImmutable();
+            }
+
+            private static Document GetReportedDocument(Diagnostic diagnostic, ImmutableDictionary<SyntaxTree, Document> treeToDocumentsMap)
+            {
+                var tree = diagnostic.Location.SourceTree;
+                if (tree != null)
+                {
+                    Document document;
+                    if (treeToDocumentsMap.TryGetValue(tree, out document))
+                    {
+                        return document;
+                    }
+                }
+
+                return null;
+            }
+
+            internal virtual async Task<ImmutableDictionary<Project, ImmutableArray<Diagnostic>>> GetProjectDiagnosticsToFixAsync(FixAllContext fixAllContext)
+            {
+                using (Logger.LogBlock(FunctionId.CodeFixes_FixAllOccurrencesComputation_Diagnostics, fixAllContext.CancellationToken))
+                {
+                    var project = fixAllContext.Project;
+                    if (project != null)
+                    {
+                        switch (fixAllContext.Scope)
+                        {
+                            case FixAllScope.Project:
+                                var diagnostics = await fixAllContext.GetProjectDiagnosticsAsync(project).ConfigureAwait(false);
+                                var kvp = SpecializedCollections.SingletonEnumerable(KeyValuePair.Create(project, diagnostics));
+                                return ImmutableDictionary.CreateRange(kvp);
+
+                            case FixAllScope.Solution:
+                                var projectsAndDiagnostics = ImmutableDictionary.CreateBuilder<Project, ImmutableArray<Diagnostic>>();
+
+                                var tasks = project.Solution.Projects.Select(async p => new
+                                {
+                                    Project = p,
+                                    Diagnostics = await fixAllContext.GetProjectDiagnosticsAsync(p).ConfigureAwait(false)
+                                }).ToArray();
+
+                                await Task.WhenAll(tasks).ConfigureAwait(false);
+
+                                foreach (var task in tasks)
+                                {
+                                    if (task.Result.Diagnostics.Any())
+                                    {
+                                        projectsAndDiagnostics[task.Result.Project] = task.Result.Diagnostics;
+                                    }
+                                }
+
+                                return projectsAndDiagnostics.ToImmutable();
+                        }
+                    }
+
+                    return ImmutableDictionary<Project, ImmutableArray<Diagnostic>>.Empty;
+                }
+            }
         }
     }
 }
