+++ src/Workspaces/Core/Portable/CodeFixes/FixAllOccurrences/FixAllContext.cs
@@ -1,12 +1,15 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeActions;
+using Microsoft.CodeAnalysis.GeneratedCodeRecognition;
+using Microsoft.CodeAnalysis.Internal.Log;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CodeFixes
@@ -392,5 +395,15 @@ namespace Microsoft.CodeAnalysis.CodeFixes
         }
 
         #endregion
+
+        internal Task<ImmutableDictionary<Document, ImmutableArray<Diagnostic>>> GetDocumentDiagnosticsToFixAsync()
+        {
+            return _diagnosticProvider.GetDocumentDiagnosticsToFixAsync(this);
+        }
+
+        internal Task<ImmutableDictionary<Project, ImmutableArray<Diagnostic>>> GetProjectDiagnosticsToFixAsync()
+        {
+            return _diagnosticProvider.GetProjectDiagnosticsToFixAsync(this);
+        }
     }
 }
